feat(open-api-page): avoid rendering nested paths and sidebar if there is only a single service
