+++ packages/open-api-page/style.module.css
@@ -33,4 +33,10 @@
     width: 0;
     flex-grow: 1;
   }
+
+  &.isSingleService {
+    max-width: 50rem;
+    margin-left: auto;
+    margin-right: auto;
+  }
 }
